+++ kitty/options/definition.py
@@ -589,15 +589,14 @@ Wayland. Use negative numbers to change scroll direction.
 '''
     )
 
-opt('pixel_scroll', 'no',
-    option_type='to_bool', ctype='bool',
-    long_text='''
-Enable per-pixel scrolling for high precision input devices (for example
-touchpads). When enabled, kitty's own scrollback can move by sub-line increments
-instead of only whole lines. This does not affect applications running inside
-the terminal (for example full-screen TUIs) that handle scrolling themselves.
-'''
-    )
+opt('pixel_scroll', 'yes', option_type='to_bool', ctype='bool', long_text='''
+Enable per-pixel scrolling, in the kitty scrollback buffer, when using high precision
+input devices (for example touchpads). When enabled, kitty's own scrollback will move
+by sub-line increments instead of only whole lines. This does not affect applications
+running inside the terminal (for example full-screen TUIs) that handle scrolling
+themselves.
+''')
+
 egr()  # }}}