+++ kitty/options/parse.py
@@ -1159,6 +1159,9 @@ class Parser:
     def paste_actions(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['paste_actions'] = paste_actions(val)
 
+    def pixel_scroll(self, val: str, ans: dict[str, typing.Any]) -> None:
+        ans['pixel_scroll'] = to_bool(val)
+
     def placement_strategy(self, val: str, ans: dict[str, typing.Any]) -> None:
         val = val.lower()
         if val not in self.choices_for_placement_strategy:
@@ -1388,9 +1391,6 @@ class Parser:
 
     def touch_scroll_multiplier(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['touch_scroll_multiplier'] = float(val)
-
-    def pixel_scroll(self, val: str, ans: dict[str, typing.Any]) -> None:
-        ans['pixel_scroll'] = to_bool(val)
 
     def transparent_background_colors(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['transparent_background_colors'] = transparent_background_colors(val)