+++ kitty/screen.c
@@ -89,6 +89,18 @@ init_overlay_line(Screen *self, index_type columns, bool keep_active) {
     return true;
 }
 
+static bool
+init_blank_line(Screen *self) {
+    const size_t sz = (sizeof(CPUCell) + sizeof(GPUCell)) * self->columns;
+    self->blank_line_buffer = PyMem_Realloc(self->blank_line_buffer, sz);
+    if (!self->blank_line_buffer) return false;
+    memset(self->blank_line_buffer, 0, sz);
+    self->blank_line = (Line){.cpu_cells=self->blank_line_buffer,
+        .gpu_cells=(GPUCell*)((CPUCell*)self->blank_line_buffer + self->columns),
+        .xnum=self->columns, .text_cache=self->text_cache};
+    return true;
+}
+
 static void deactivate_overlay_line(Screen *self);
 static void update_overlay_position(Screen *self);
 static void render_overlay_line(Screen *self, Line *line, FONTS_DATA_HANDLE fonts_data);
@@ -162,15 +174,7 @@ new_screen_object(PyTypeObject *type, PyObject *args, PyObject UNUSED *kwds) {
         init_tabstops(self->alt_tabstops, self->columns);
         self->key_encoding_flags = self->main_key_encoding_flags;
         if (!init_overlay_line(self, self->columns, false)) { Py_CLEAR(self); return NULL; }
-        self->blank_line_cpu = PyMem_Calloc(self->columns, sizeof(CPUCell));
-        self->blank_line_gpu = PyMem_Calloc(self->columns, sizeof(GPUCell));
-        if (!self->blank_line_cpu || !self->blank_line_gpu) { Py_CLEAR(self); return NULL; }
-        self->blank_line.cpu_cells = self->blank_line_cpu;
-        self->blank_line.gpu_cells = self->blank_line_gpu;
-        self->blank_line.xnum = self->columns;
-        self->blank_line.ynum = 0;
-        self->blank_line.attrs.val = 0;
-        self->blank_line.text_cache = self->text_cache;
+        if (!init_blank_line(self)) { Py_CLEAR(self); return NULL; }
         self->hyperlink_pool = alloc_hyperlink_pool();
         if (!self->hyperlink_pool) { Py_CLEAR(self); return PyErr_NoMemory(); }
         self->as_ansi_buf.hyperlink_pool = self->hyperlink_pool;
@@ -557,19 +561,8 @@ screen_resize(Screen *self, unsigned int lines, unsigned int columns) {
     which.is_beyond_content = num_content_lines_before > 0 && self->cursor->y >= num_content_lines_before; \
     which.num_content_lines = num_content_lines_after; \
 }
-    // Resize overlay line
-    if (!init_overlay_line(self, columns, true)) return false;
-    PyMem_Free(self->blank_line_cpu);
-    PyMem_Free(self->blank_line_gpu);
-    self->blank_line_cpu = PyMem_Calloc(columns, sizeof(CPUCell));
-    self->blank_line_gpu = PyMem_Calloc(columns, sizeof(GPUCell));
-    if (!self->blank_line_cpu || !self->blank_line_gpu) return false;
-    self->blank_line.cpu_cells = self->blank_line_cpu;
-    self->blank_line.gpu_cells = self->blank_line_gpu;
-    self->blank_line.xnum = columns;
-    self->blank_line.ynum = 0;
-    self->blank_line.attrs.val = 0;
-    self->blank_line.text_cache = self->text_cache;
+    // Resize overlay and blank lines
+    if (!init_overlay_line(self, columns, true) || !init_blank_line(self)) return false;
 
     // Resize main linebuf
     RAII_PyObject(prompt_copy, NULL);
@@ -683,6 +676,7 @@ dealloc(Screen* self) {
     PyMem_Free(self->overlay_line.gpu_cells);
     PyMem_Free(self->overlay_line.original_line.cpu_cells);
     PyMem_Free(self->overlay_line.original_line.gpu_cells);
+    PyMem_Free(self->blank_line_buffer);
     Py_CLEAR(self->overlay_line.overlay_text);
     PyMem_Free(self->main_tabstops);
     Py_CLEAR(self->paused_rendering.linebuf);