+++ kitty/screen.h
@@ -109,9 +109,7 @@ typedef struct {
     double pixel_scroll_offset_y;
     CellPixelSize cell_size;
     OverlayLine overlay_line;
-    Line blank_line;
-    CPUCell *blank_line_cpu;
-    GPUCell *blank_line_gpu;
+    Line blank_line; void *blank_line_buffer;
     id_type window_id;
     Selections selections, url_ranges;
     struct {